Add documentation for cookiebar and move cookie-bar.js into it's own file that's exported by index.js
